Refactor InsertSlice(Async) to LoadStore
